Extract KiSSDoll module
